requirements.txt
streamlit==1.32.0
PyMuPDF==1.23.26
requests==2.31.0
python-dotenv==1.0.1
pandas==2.2.1
Pillow==10.2.0
python-magic-bin==0.4.14 